백그라운드 디버깅 체크 시작/종료 수동 처리 추가

+++ CheckDebug/src/main/cpp/check_debug_util.cpp
@@ -3,10 +3,14 @@
 #include <unistd.h>
 #include <android/log.h>
 #include <sys/ptrace.h>
+#include <sys/stat.h>
 #include <sys/wait.h>
 #include <pthread.h>
 #include <chrono>
 #include <thread>
+#include <vector>
+#include <string>
+#include <fstream>
 
 #define  LOG_TAG    "CheckDebug"
 #define  LOGUNK(...)  __android_log_print(ANDROID_LOG_UNKNOWN,LOG_TAG,__VA_ARGS__)
@@ -19,9 +23,103 @@
 #define  LOGF(...)  __android_log_print(ANDROID_FATAL_ERROR,LOG_TAG,__VA_ARGS__)
 #define  LOGS(...)  __android_log_print(ANDROID_SILENT_ERROR,LOG_TAG,__VA_ARGS__)
 
+bool fileExists(const std::string &filePath) {
+    struct stat info;
+    return ( stat( filePath.c_str(), &info ) == 0 );
+}
+
+bool checkRootPath(const std::string &path) {
+    struct stat info;
+    if (stat(path.c_str(), &info) != 0)
+        return false;
+    return (info.st_mode & S_IFDIR) != 0;
+}
+
+std::string findExecutable(const std::string &command) {
+    std::string path;
+    std::string cmd = "which " + command + " 2>/dev/null";
+
+    FILE *pipe = popen(cmd.c_str(), "r");
+    if (pipe != nullptr) {
+        char buffer[128];
+        if (fgets(buffer, sizeof(buffer), pipe) != nullptr) {
+            path = buffer;
+            // Remove trailing newline character
+            if (!path.empty() && path.back() == '\n') {
+                path.pop_back();
+            }
+        }
+        pclose(pipe);
+    }
+
+    return path;
+}
+
+static bool isRooted() {
+    // Check for common rooted paths
+    std::vector<std::string> rootedPaths = {
+            "/sbin/su",
+            "/system/su",
+            "/system/bin/su",
+            "/system/sbin/su",
+            "/system/xbin/su",
+            "/system/xbin/mu",
+            "/system/bin/.ext/.su",
+            "/system/usr/su-backup",
+            "/data/data/com.noshufou.android.su",
+            "/system/app/Superuser.apk",
+            "/system/app/su.apk",
+            "/system/bin/.ext",
+            "/system/xbin/.ext",
+            "/data/local/xbin/su",
+            "/data/local/bin/su",
+            "/system/sd/xbin/su",
+            "/system/bin/failsafe/su",
+            "/data/local/su",
+            "/su/bin/su"
+    };
+
+    for (const std::string &path: rootedPaths) {
+        if (fileExists(path) || checkRootPath(path)) {
+            LOGD("rooted path: %s", (path.c_str()));
+            return true;
+        }
+    }
+
+    // Check for installed root management apps
+    std::vector<std::string> rootApps = {
+            "com.noshufou.android.su",
+            "eu.chainfire.supersu",
+            "com.koushikdutta.superuser",
+            "com.thirdparty.superuser",
+            "com.topjohnwu.magisk",
+            "com.playground.rooting"
+    };
+
+    for (const std::string &packageName: rootApps) {
+        std::string cmd = std::string("pm list packages ").append(packageName) + " | grep " + packageName;
+        if (system(cmd.c_str()) == 0) {
+            LOGD("rooted package: %s", cmd.c_str());
+            return true;
+        }
+    }
+
+    return false;
+}
+
+bool isCheckExecution() {
+    std::string suPath = findExecutable("su");
+    if (!suPath.empty()) {
+        return true;
+    }
+
+    return false;
+}
+
+
 extern "C"
 JNIEXPORT jboolean JNICALL
-Java_kr_ds_util_CheckDebugNativeLib_isDebugToolTracerPid(JNIEnv *env, jobject thiz) {
+Java_kr_ds_util_CheckDebugNativeLib_isDebugToolTracerPid(JNIEnv *, jobject) {
     int TPid;
     char buf[512];
     const char *str = "TracerPid:";
@@ -50,7 +148,7 @@ Java_kr_ds_util_CheckDebugNativeLib_isDebugToolTracerPid(JNIEnv *env, jobject th
 }
 extern "C"
 JNIEXPORT jboolean JNICALL
-Java_kr_ds_util_CheckDebugNativeLib_isDebugToolCmdLine(JNIEnv *env, jobject thiz) {
+Java_kr_ds_util_CheckDebugNativeLib_isDebugToolCmdLine(JNIEnv *, jobject) {
     char filePath[32], fileRead[128];
     FILE *file;
 
@@ -78,11 +176,11 @@ Java_kr_ds_util_CheckDebugNativeLib_isDebugToolCmdLine(JNIEnv *env, jobject thiz
     return false;
 }
 
-const int version = 2;
+const int version = 3;
 
 extern "C"
 JNIEXPORT jint JNICALL
-Java_kr_ds_util_CheckDebugNativeLib_version(JNIEnv *env, jobject thiz) {
+Java_kr_ds_util_CheckDebugNativeLib_version(JNIEnv *env, jobject) {
     LOGD("version = %d", version);
     return version;
 }
@@ -143,9 +241,16 @@ void anti_debug_by_preemption_ptrace() {
 
         /* Start the monitoring thread */
         LOGD("anti_debug_by_preemption_ptrace Start the monitoring thread");
-        pthread_create(&t, NULL, reinterpret_cast<void *(*)(void *)>(monitor_pid), (void *) NULL);
+        pthread_create(&t, nullptr, reinterpret_cast<void *(*)(void *)>(monitor_pid),
+                       (void *) nullptr);
     }
 }
+
+static pthread_key_t g_key;
+static JavaVM *pJavaVm;
+static jobject callback_object = nullptr;
+
+static JNIEnv *getEnv();
 
 void be_attached_check() {
 
@@ -174,12 +279,18 @@ void be_attached_check() {
                     LOGV("%s", line);
 
                     if (statue != 0) {
-
-                        LOGD("be attached !! kill %d", pid);
-
                         fclose(fd);
 
-                        int ret = kill(pid, SIGKILL);
+                        if (callback_object) {
+                            JNIEnv *_env = getEnv();
+                            jclass clazz = _env->GetObjectClass(callback_object);
+                            jmethodID method = _env->GetMethodID(clazz, "onDebug", "()V");
+                            _env->CallVoidMethod(callback_object, method);
+                        } else {
+                            LOGD("be attached !! kill %d", pid);
+                            int ret = kill(pid, SIGKILL);
+                        }
+
 
                     }
 
@@ -201,9 +312,11 @@ void be_attached_check() {
     }
 }
 
+static bool isAliveAntiDebugByProcessStatus = false;
+
 void thread_task(int n) {
 
-    while (true) {
+    while (isAliveAntiDebugByProcessStatus) {
 
         LOGV("start be_attached_check...");
 
@@ -217,7 +330,7 @@ void thread_task(int n) {
 
 void anti_debug_by_process_status() {
 
-    LOGD("call anti_debug_by_preemption_ptrace by process status ......");
+    LOGD("called anti_debug_by_process_status");
 
     auto checkThread = std::thread(thread_task, 1);
 
@@ -225,11 +338,67 @@ void anti_debug_by_process_status() {
 
 }
 
+JavaVM *getJavaVM() {
+    pthread_t thisThread = pthread_self();
+    LOGD("getJavaVM(), pthread_self() = %ld", thisThread);
+    return pJavaVm;
+}
+
+static void detachCurrentThread(void *a) {
+    getJavaVM()->DetachCurrentThread();
+}
+
+void setJavaVM(JavaVM *javaVM) {
+    pthread_t thisThread = pthread_self();
+    LOGD("setJavaVM(%p), pthread_self() = %ld", javaVM, thisThread);
+    pJavaVm = javaVM;
+
+    pthread_key_create(&g_key, detachCurrentThread);
+}
+
+JNIEnv *cacheEnv(JavaVM *jvm) {
+    JNIEnv *_env = nullptr;
+    // get jni environment
+    jint ret = jvm->GetEnv((void **) &_env, JNI_VERSION_1_6);
+
+    switch (ret) {
+        case JNI_OK :
+            // Success!
+            pthread_setspecific(g_key, _env);
+            return _env;
+
+        case JNI_EDETACHED :
+            // Thread not attached
+            if (jvm->AttachCurrentThread(&_env, nullptr) < 0) {
+                LOGE("Failed to get the environment using AttachCurrentThread()");
+
+                return nullptr;
+            } else {
+                // Success : Attached and obtained JNIEnv!
+                pthread_setspecific(g_key, _env);
+                return _env;
+            }
+
+        case JNI_EVERSION :
+            // Cannot recover from this error
+            LOGE("JNI interface version 1.6 not supported");
+        default :
+            LOGE("Failed to get the environment using GetEnv()");
+            return nullptr;
+    }
+}
+
+JNIEnv *getEnv() {
+    auto *_env = (JNIEnv *) pthread_getspecific(g_key);
+    if (_env == nullptr)
+        _env = cacheEnv(pJavaVm);
+    return _env;
+}
+
 jint JNI_OnLoad(JavaVM *vm, void *reserved) {
-
-    anti_debug_by_process_status();
-
     JNIEnv *env;
+
+    setJavaVM(vm);
 
     if (vm->GetEnv(reinterpret_cast<void **>(&env), JNI_VERSION_1_6) != JNI_OK) {
 
@@ -243,6 +412,70 @@ jint JNI_OnLoad(JavaVM *vm, void *reserved) {
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_kr_ds_util_CheckDebugNativeLib_antiDebug(JNIEnv *env, jobject thiz) {
+Java_kr_ds_util_CheckDebugNativeLib_antiDebug(JNIEnv *env, jobject) {
     anti_debug_by_preemption_ptrace();
+}
+
+
+
+extern "C"
+JNIEXPORT void JNICALL
+Java_kr_ds_util_CheckDebugNativeLib_00024Companion_startAntiDebugOnBackground(JNIEnv *env,
+                                                                              jobject,
+                                                                              jobject callback) {
+    if (isAliveAntiDebugByProcessStatus) {
+        return;
+    }
+    LOGD("startAntiDebugOnBackground %p", callback);
+    callback_object = env->NewGlobalRef(callback);
+
+    isAliveAntiDebugByProcessStatus = true;
+    anti_debug_by_process_status();
+}
+
+extern "C"
+JNIEXPORT void JNICALL
+Java_kr_ds_util_CheckDebugNativeLib_00024Companion_stopAntiDebugOnBackground(JNIEnv *env,
+                                                                             jobject) {
+    isAliveAntiDebugByProcessStatus = false;
+    if (callback_object) {
+        env->DeleteGlobalRef(callback_object);
+        callback_object = nullptr;
+    }
+}
+
+static bool isDebuggable() {
+    std::string getpropCmd = "getprop ro.debuggable";
+    std::string debuggableValue;
+    FILE *pipe = popen(getpropCmd.c_str(), "r");
+    if (pipe != nullptr) {
+        char buffer[128];
+        if (fgets(buffer, sizeof(buffer), pipe) != nullptr) {
+            debuggableValue = buffer;
+
+            if (!debuggableValue.empty() && debuggableValue.back() == '\n') {
+                debuggableValue.pop_back();
+            }
+        }
+        pclose(pipe);
+    }
+
+    LOGD("ro.debuggable: %s", debuggableValue.c_str());
+    return (debuggableValue == "1");
+}
+
+extern "C"
+JNIEXPORT jboolean JNICALL
+Java_kr_ds_util_CheckDebugNativeLib_isDebugEnabled(JNIEnv *env, jobject) {
+    return isDebuggable() ? JNI_TRUE : JNI_FALSE;
+}
+extern "C"
+JNIEXPORT jboolean JNICALL
+Java_kr_ds_util_CheckDebugNativeLib_isRootingByFileExistence(JNIEnv *env, jobject) {
+    return isRooted() ? JNI_TRUE : JNI_FALSE;
+}
+extern "C"
+JNIEXPORT jboolean JNICALL
+Java_kr_ds_util_CheckDebugNativeLib_isRootingByExecCmd(JNIEnv *env, jobject) {
+    return isCheckExecution() ? JNI_TRUE : JNI_FALSE;
 }
